hockeyapp for release 1
