+++ Habitica/AndroidManifest.xml
@@ -76,9 +76,6 @@
         </activity>
 
         <meta-data android:name="com.facebook.sdk.ApplicationId" android:value="@string/facebook_app_id"/>
-
-        <meta-data android:name="net.hockeyapp.android.appIdentifier" android:value="ee347e558b7e49f7bc5e54c102efab66" />
-
         <meta-data
                 android:name="com.facebook.sdk.AutoLogAppEventsEnabled"
                 android:value="false"/>
