+++ Habitica/src/main/java/com/habitrpg/android/habitica/ui/activities/MainActivity.java
@@ -25,6 +25,7 @@ import android.support.v4.content.ContextCompat;
 import android.support.v4.content.FileProvider;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.Toolbar;
+import android.util.Log;
 import android.util.TypedValue;
 import android.view.KeyEvent;
 import android.view.View;
@@ -105,11 +106,6 @@ import com.mikepenz.materialdrawer.model.PrimaryDrawerItem;
 import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem;
 import com.mikepenz.materialdrawer.model.interfaces.IProfile;
 import com.roughike.bottombar.BottomBar;
-
-import net.hockeyapp.android.CrashManager;
-import net.hockeyapp.android.FeedbackManager;
-import net.hockeyapp.android.UpdateManager;
-import net.hockeyapp.android.metrics.MetricsManager;
 
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
@@ -278,18 +274,6 @@ public class MainActivity extends BaseActivity implements TutorialView.OnTutoria
                 }, RxErrorHandler.handleEmptyError());
 
         EventBus.getDefault().register(this);
-
-        MetricsManager.register(getApplication());
-        checkForUpdates();
-    }
-
-    private void checkForUpdates() {
-        // Remove this for store builds!
-        UpdateManager.register(this);
-    }
-
-    private void unregisterManagers() {
-        UpdateManager.unregister();
     }
 
     public int getStatusBarHeight() {
@@ -336,9 +320,6 @@ public class MainActivity extends BaseActivity implements TutorialView.OnTutoria
                 drawer.setSelectionAtPosition(this.sharedPreferences.getInt("lastActivePosition", 1));
             }
         }
-
-        CrashManager.register(this);
-        FeedbackManager.register(this);
     }
 
     @Override
@@ -348,7 +329,6 @@ public class MainActivity extends BaseActivity implements TutorialView.OnTutoria
         updateWidget(DailiesWidgetProvider.class);
         updateWidget(HabitButtonWidgetProvider.class);
         super.onPause();
-        unregisterManagers();
     }
 
     private void updateWidget(Class widgetClass) {
@@ -570,7 +550,6 @@ public class MainActivity extends BaseActivity implements TutorialView.OnTutoria
         tagRepository.close();
         inventoryRepository.close();
         super.onDestroy();
-        unregisterManagers();
     }
 
     @Subscribe
